compiles without warnings

+++ include/game_show.h
@@ -17,7 +17,7 @@ using namespace std;
 #include "coil.h"
 //#include "lamp_matrix_test.h"
 
-#define SIZEOF(n) (sizeof(n) / sizeof(n[0]))
+#define SIZEOF(n) ((int)(sizeof(n) / sizeof(n[0])))
 #define pohoNULL (0)
